No commit message

+++ install-go.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-. $(realpath $(dirname $0))/install-helper.sh
+set -euo pipefail
 
 os=linux
 if [[ $(uname) == Darwin ]]; then
@@ -7,19 +7,16 @@ if [[ $(uname) == Darwin ]]; then
 fi
 ver=$(wget -qO- https://golang.org/dl/ | grep -Po '(?<=/dl/go)[0-9.]+(?=\.'$os'-amd64\.tar\.gz)' | head -n1)
 url=https://golang.org/dl/go$ver.$os-amd64.tar.gz
-exe=go
-local_ver() {
-	go version | grep -Po '(?<=go)[0-9.]+'
-}
-install_func() {
-	local tmp=$(mktemp_track)
-	echo "Downloading"
-	wget -qO $tmp $url
-	echo "Deleting old goroot"
-	rm -rf ~/.goroot
-	echo "Extracting"
-	tar -xf $tmp -C $HOME --transform=s/^go/.goroot/
-	echo "Installed"
-}
 
-helper
+echo "Installed:     $(go version 2>/dev/null | grep -Po '(?<=go)[0-9.]+' || echo "Not installed")"
+echo "Remote latest: $ver"
+
+if [[ ${1:-} == -n ]]; then
+	exit
+fi
+
+tmp=$(mktemp)
+wget -qO $tmp $url
+rm -rf ~/.goroot
+tar -xf $tmp -C $HOME --transform=s/^go/.goroot/
+rm -f $tmp
